Moving react-next TeachingBubble to react-internal
