--- apps/perf-test/src/scenarios/TeachingBubbleNext.tsx
@@ -1,6 +1,0 @@
-import * as React from 'react';
-import { TeachingBubble } from '@fluentui/react-next';
-
-const Scenario = () => <TeachingBubble />;
-
-export default Scenario;
