--- packages/react-examples/src/react-next/TeachingBubble/TeachingBubble.Illustration.Example.tsx
@@ -1,47 +1,0 @@
-import * as React from 'react';
-import { IImageProps } from '@fluentui/react-next/lib/Image';
-import { DefaultButton, IButtonProps } from '@fluentui/react-next/lib/compat/Button';
-import { TeachingBubble } from '@fluentui/react-next/lib/TeachingBubble';
-import { useBoolean } from '@uifabric/react-hooks';
-
-const exampleImageProps: IImageProps = { src: 'http://placehold.it/364x180', alt: 'Example placeholder image' };
-
-const examplePrimaryButtonProps: IButtonProps = {
-  children: 'Try it out',
-};
-
-export const TeachingBubbleIllustrationExample: React.FunctionComponent = () => {
-  const [teachingBubbleVisible, { toggle: toggleTeachingBubbleVisible }] = useBoolean(false);
-
-  const exampleSecondaryButtonProps: IButtonProps = React.useMemo(
-    () => ({
-      children: 'Maybe later',
-      onClick: toggleTeachingBubbleVisible,
-    }),
-    [toggleTeachingBubbleVisible],
-  );
-
-  return (
-    <div>
-      <DefaultButton
-        id="targetButton"
-        onClick={toggleTeachingBubbleVisible}
-        text={teachingBubbleVisible ? 'Hide TeachingBubble' : 'Show TeachingBubble'}
-      />
-
-      {teachingBubbleVisible && (
-        <TeachingBubble
-          target="#targetButton"
-          illustrationImage={exampleImageProps}
-          primaryButtonProps={examplePrimaryButtonProps}
-          secondaryButtonProps={exampleSecondaryButtonProps}
-          onDismiss={toggleTeachingBubbleVisible}
-          headline="Discover what’s trending around you"
-        >
-          Lorem ipsum dolor sit amet, consectetur adipisicing elit. Facere, nulla, ipsum? Molestiae quis aliquam magni
-          harum non?
-        </TeachingBubble>
-      )}
-    </div>
-  );
-};
