--- packages/react-examples/src/react-next/TeachingBubble/TeachingBubble.MultiStep.Example.tsx
@@ -1,43 +1,0 @@
-import * as React from 'react';
-import { DefaultButton, IButtonProps } from '@fluentui/react-next/lib/compat/Button';
-import { TeachingBubble } from '@fluentui/react-next/lib/TeachingBubble';
-import { useBoolean } from '@uifabric/react-hooks';
-
-const examplePrimaryButtonProps: IButtonProps = {
-  children: 'Next',
-};
-export const TeachingBubbleMultiStepExample: React.FunctionComponent = () => {
-  const [teachingBubbleVisible, { toggle: toggleTeachingBubbleVisible }] = useBoolean(false);
-
-  const exampleSecondaryButtonProps: IButtonProps = React.useMemo(
-    () => ({
-      children: 'Previous',
-      onClick: toggleTeachingBubbleVisible,
-    }),
-    [toggleTeachingBubbleVisible],
-  );
-
-  return (
-    <div>
-      <DefaultButton
-        id="targetButton"
-        onClick={toggleTeachingBubbleVisible}
-        text={teachingBubbleVisible ? 'Hide TeachingBubble' : 'Show TeachingBubble'}
-      />
-
-      {teachingBubbleVisible && (
-        <TeachingBubble
-          target="#targetButton"
-          primaryButtonProps={examplePrimaryButtonProps}
-          secondaryButtonProps={exampleSecondaryButtonProps}
-          onDismiss={toggleTeachingBubbleVisible}
-          footerContent="2 of 3"
-          headline="Discover what’s trending around you"
-        >
-          Lorem ipsum dolor sit amet, consectetur adipisicing elit. Facere, nulla, ipsum? Molestiae quis aliquam magni
-          harum non?
-        </TeachingBubble>
-      )}
-    </div>
-  );
-};
