--- packages/react-examples/src/react-next/TeachingBubble/TeachingBubble.Wide.Example.tsx
@@ -1,34 +1,0 @@
-import * as React from 'react';
-import { DefaultButton } from '@fluentui/react-next/lib/compat/Button';
-import { TeachingBubble } from '@fluentui/react-next/lib/TeachingBubble';
-import { DirectionalHint } from '@fluentui/react-next';
-import { useBoolean } from '@uifabric/react-hooks';
-
-export const TeachingBubbleWideExample: React.FunctionComponent = () => {
-  const [teachingBubbleVisible, { toggle: toggleTeachingBubbleVisible }] = useBoolean(false);
-
-  return (
-    <div>
-      <DefaultButton
-        id="targetButton"
-        onClick={toggleTeachingBubbleVisible}
-        text={teachingBubbleVisible ? 'Hide TeachingBubble' : 'Show TeachingBubble'}
-      />
-
-      {teachingBubbleVisible && (
-        <TeachingBubble
-          calloutProps={{ directionalHint: DirectionalHint.bottomCenter }}
-          target="#targetButton"
-          isWide={true}
-          hasCloseButton={true}
-          closeButtonAriaLabel="Close"
-          onDismiss={toggleTeachingBubbleVisible}
-          headline="Discover what’s trending around you"
-        >
-          Lorem ipsum dolor sit amet, consectetur adipisicing elit. Facere, nulla, ipsum? Molestiae quis aliquam magni
-          harum non?
-        </TeachingBubble>
-      )}
-    </div>
-  );
-};
