--- packages/react-examples/src/react-next/TeachingBubble/docs/TeachingBubbleBestPractices.md
@@ -1,27 +1,0 @@
-### Layout
-
-- Teaching bubbles can be used in sequence to walk people through complex, multistep interactions only. And only show one teaching bubble at a time.
-- A maximum of no more than 3 sequenced teaching bubbles should be used in a single experience.
-- Sequenced teaching bubbles should have “x of y” text to indicate progress through the sequence. For example, the first teaching bubble in a sequence might be “1 of 3”.
-- Always place the primary button on the left, the secondary button just to the right of it.
-- Show only one primary button that inherits theme color at rest state. In the event there are more than two buttons with equal priority, all buttons should have neutral backgrounds.
-
-### Content
-
-A teaching bubble should include:
-
-#### Title
-
-You have 25 characters (including spaces) to draw people in and get them interested. Limit to one line of text, and use sentence casing (capitalize only the first word and any proper nouns) with no punctuation.
-
-#### Body copy
-
-Lead with why the feature is useful rather than describe what it is. What does it make possible? Keep it within 120 characters (including spaces).
-
-#### Action buttons
-
-Limit button labels to 15 characters (including spaces). Provide people with an explicit action to dismiss the teaching bubble, such as “Got it”. Include a secondary button to give people the option to take action, such as “Show me” or “Edit settings”. When two buttons are needed, make the call-to-action button the primary button and the dismissal button (such as “No thanks”) the secondary button. Use sentence casing (capitalize only the first word and any proper nouns) with no punctuation. On a sequenced teaching bubble, use "Next" for the action button label and "Got it" for the final sequenced teaching bubble action button with text that closes the experience.
-
-#### Link (Optional)
-
-If there are additional steps people need to know about, or helpful information they may want to read, consider linking to a help article. Typically, these links are labeled “Learn more” with no punctuation at the end.
