--- packages/react-examples/src/react-next/TeachingBubble/docs/TeachingBubbleOverview.md
@@ -1,3 +1,0 @@
-A teaching bubble (`TeachingBubble`) brings attention to a new or important feature, with the goal of increasing engagement with the feature. A teaching bubble typically follows a coach mark.
-
-Use teaching bubbles sparingly. Consider how frequently people will see it, and how many they’ll see across their entire experience.
