+++ packages/react-internal/src/components/TeachingBubble/TeachingBubble.base.tsx
@@ -1,7 +1,8 @@
 import * as React from 'react';
-import { initializeComponentRef, classNamesFunction } from '../../Utilities';
+import { classNamesFunction } from '../../Utilities';
 import { TeachingBubbleContent } from './TeachingBubbleContent';
 import {
+  ITeachingBubble,
   ITeachingBubbleProps,
   ITeachingBubbleStyleProps,
   ITeachingBubbleStyles,
@@ -9,95 +10,79 @@ import {
 } from './TeachingBubble.types';
 import { Callout, ICalloutProps } from '../../Callout';
 import { DirectionalHint } from '../../common/DirectionalHint';
+import { useMergedRefs } from '@uifabric/react-hooks';
+
+const COMPONENT_NAME = 'TeachingBubble';
+const defaultCalloutProps: ICalloutProps = {
+  beakWidth: 16,
+  gapSpace: 0,
+  setInitialFocus: true,
+  doNotLayer: false,
+  directionalHint: DirectionalHint.rightCenter,
+};
 
 const getClassNames = classNamesFunction<ITeachingBubbleStyleProps, ITeachingBubbleStyles>();
+const useComponentRef = (
+  componentRef: React.Ref<ITeachingBubble> | undefined,
+  rootElementRef: React.RefObject<HTMLDivElement>,
+) => {
+  React.useImperativeHandle(
+    componentRef,
+    () => ({
+      focus: () => rootElementRef.current?.focus(),
+    }),
+    [rootElementRef],
+  );
+};
 
-export interface ITeachingBubbleState {
-  isTeachingBubbleVisible?: boolean;
-}
-
-export class TeachingBubbleBase extends React.Component<ITeachingBubbleProps, ITeachingBubbleState> {
-  public static defaultProps = {
-    /**
-     * Default calloutProps is deprecated in favor of private `_defaultCalloutProps`.
-     * Remove in next release.
-     * @deprecated In favor of private `_defaultCalloutProps`.
-     */
+export const TeachingBubbleBase: React.FunctionComponent<ITeachingBubbleProps> = React.forwardRef<
+  HTMLDivElement,
+  ITeachingBubbleProps
+>((props, forwardedRef) => {
+  const rootElementRef = React.useRef<HTMLDivElement>(null);
+  const mergedRootRef = useMergedRefs(rootElementRef, forwardedRef);
+  const {
+    calloutProps: setCalloutProps,
     // eslint-disable-next-line deprecation/deprecation
-    calloutProps: {
-      beakWidth: 16,
-      gapSpace: 0,
-      setInitialFocus: true,
-      doNotLayer: false,
-      directionalHint: DirectionalHint.rightCenter,
-    },
+    targetElement,
+    onDismiss,
+    // eslint-disable-next-line deprecation/deprecation
+    hasCloseButton = props.hasCloseIcon,
+    isWide,
+    styles,
+    theme,
+    target,
+  } = props;
+
+  const calloutProps = { ...defaultCalloutProps, ...setCalloutProps };
+
+  const stylesProps: ITeachingBubbleStyleProps = {
+    theme: theme!,
+    isWide,
+    calloutProps: { ...calloutProps, theme: calloutProps.theme! },
+    hasCloseButton,
   };
 
-  public rootElement = React.createRef<HTMLDivElement>();
-  private _defaultCalloutProps: ICalloutProps;
+  const classNames = getClassNames(styles, stylesProps);
+  const calloutStyles = classNames.subComponentStyles
+    ? (classNames.subComponentStyles as ITeachingBubbleSubComponentStyles).callout
+    : undefined;
 
-  // Constructor
-  constructor(props: ITeachingBubbleProps) {
-    super(props);
+  useComponentRef(props.componentRef, rootElementRef);
 
-    initializeComponentRef(this);
-    this.state = {};
-
-    this._defaultCalloutProps = {
-      beakWidth: 16,
-      gapSpace: 0,
-      setInitialFocus: true,
-      doNotLayer: false,
-      directionalHint: DirectionalHint.rightCenter,
-    };
-  }
-
-  public focus(): void {
-    if (this.rootElement.current) {
-      this.rootElement.current.focus();
-    }
-  }
-
-  public render(): JSX.Element {
-    const {
-      calloutProps: setCalloutProps,
-      // eslint-disable-next-line deprecation/deprecation
-      targetElement,
-      onDismiss,
-      // Default to deprecated value if provided.
-      // eslint-disable-next-line deprecation/deprecation
-      hasCloseButton = this.props.hasCloseIcon,
-      isWide,
-      styles,
-      theme,
-      target,
-    } = this.props;
-    const calloutProps = { ...this._defaultCalloutProps, ...setCalloutProps };
-    const stylesProps: ITeachingBubbleStyleProps = {
-      theme: theme!,
-      isWide,
-      calloutProps: { ...calloutProps, theme: calloutProps.theme! },
-      hasCloseButton,
-    };
-
-    const classNames = getClassNames(styles, stylesProps);
-    const calloutStyles = classNames.subComponentStyles
-      ? (classNames.subComponentStyles as ITeachingBubbleSubComponentStyles).callout
-      : undefined;
-
-    return (
-      <Callout
-        target={target || targetElement}
-        onDismiss={onDismiss}
-        {...calloutProps}
-        className={classNames.root}
-        styles={calloutStyles}
-        hideOverflow
-      >
-        <div ref={this.rootElement}>
-          <TeachingBubbleContent {...this.props} />
-        </div>
-      </Callout>
-    );
-  }
-}
+  return (
+    <Callout
+      target={target || targetElement}
+      onDismiss={onDismiss}
+      {...calloutProps}
+      className={classNames.root}
+      styles={calloutStyles}
+      hideOverflow
+    >
+      <div ref={mergedRootRef}>
+        <TeachingBubbleContent {...props} />
+      </div>
+    </Callout>
+  );
+});
+TeachingBubbleBase.displayName = COMPONENT_NAME;
