--- packages/react-next/src/components/TeachingBubble/TeachingBubble.deprecated.test.tsx
@@ -1,20 +1,0 @@
-import * as React from 'react';
-import * as renderer from 'react-test-renderer';
-import { TeachingBubbleContent } from './TeachingBubbleContent';
-
-describe('TeachingBubble', () => {
-  it('renders renders with hasCloseIcon which is deprecated', () => {
-    const componentContent = renderer.create(
-      <TeachingBubbleContent
-        headline="Test Title"
-        hasCloseIcon={true}
-        primaryButtonProps={{ children: 'Test Primary Button', className: 'primary-className' }}
-        secondaryButtonProps={{ children: 'Test Secondary Button', className: 'secondary-className' }}
-      >
-        Content
-      </TeachingBubbleContent>,
-    );
-    const treeContent = componentContent.toJSON();
-    expect(treeContent).toMatchSnapshot();
-  });
-});
