--- packages/react-next/src/components/TeachingBubble/index.ts
@@ -1,5 +1,0 @@
-export * from './TeachingBubble';
-export * from './TeachingBubble.base';
-export * from './TeachingBubble.types';
-export * from './TeachingBubbleContent';
-export * from './TeachingBubbleContent.base';
